requirements.txt
gymnasium>=0.29.0
numpy>=1.24.0
torch>=2.0.0
pygame>=2.5.0
pyyaml>=6.0
tqdm>=4.65.0
